implemented confirm card payment

+++ src/Pages/Dashboard/MyCart/MyCart.jsx
@@ -48,7 +48,7 @@ const MyCart = () => {
             <div className="bg-white p-10 rounded-xl">
                 <div className="flex gap-10 justify-evenly font-semibold mb-5">
                     <h3 className="text-3xl">Total Orders: {cart.length}</h3>
-                    <h3 className="text-3xl">Total Price: ${parseInt(total).toFixed(2)}</h3>
+                    <h3 className="text-3xl">Total Price: ${parseFloat(total).toFixed(2)}</h3>
                     <Link to='/dashboard/reservation'><button className="btn btn-warning btn-sm">Pay</button></Link>
                 </div>
                 <div className="overflow-x-auto w-full">
